Update QuadTree.cpp
minor type fix

+++ QuadTree/QuadTree.cpp
@@ -18,7 +18,7 @@ class QT
 
     std::vector<MyVec*> data;
 public:
-    QT(int h, int x0, int x1, int y0, int y1) {
+    QT(int h, float x0, float x1, float y0, float y1) {
         this->prof = h; // profundidad
         this->x0 = x0;
         this->y0 = y0;
